1) Added xcode project files 2) Fixed static checker issues 3) added .dat as export format - recognizable by Octave

+++ inc/sp_mem.h
@@ -30,5 +30,5 @@ void spfree(void* ptr);
  * appropriate memory block pointed by src
  * Returns the pointer to the newly allocated block
  */
-void* memdup(const void* src, int bytes);
+void* memdup(const void* src, size_t bytes);
